update: move resolve.rs to key_effect.rs
